requirements.txt
httpx
bs4
Jinja2
lxml
python-slugify